Better approach. Write out shims.
Every time wrap() is called, write out bash shims to act as the
first 'node' or 'iojs' in the path.  A bash shim is written, along
with a '_env' file and a '_args' file, in a temp dir.

On unwrap() or process exit, we delete that folder.

No parsing anywhere.