Add special handling for spawning `npm`
As laid out in bcoe/nyc#190, spawn-wrap’ing `npm` does currently
not work when using the `npm` binary that comes bundled with Node.js.

This adds special handler code for the case that `npm` is the
program that should be executed, looking for the `npm` executable
in PATH using `which` and invoking the shim with it as an argument.