Fixups for the recent shebang changes
`options.args[0]` contains the executable name, so it should not
be included in the newly built `options.args` array when a
shebang is found, because the script’s file name is already passed as
`resolved`.

Also set the local `file` variable to the new executable so
spawn-wrap does not attempt to double-fix in the case of
encountering `file === 'npm'`.