Massive refactor
- Move all the is/munge functionality into functions
- Track the originalNode in the settings and environment
- Put XXX comments in a few places where it's clear that a bug is
  lurking.

This is a step closer to fixing #38, but it's still not there yet.

The right answer probably will NOT have the script passing the original
requested node through to the shim, but rather injecting the shim into
the args array after the node binary.  This will require that the shim's
runMain function know to strip itself off of the process.argv array
(along with any other stuff added there).

Next step in the refactor will be to pull the is/munge functions into
separate modules that can be independently tested.  I'd also like to add
some tests to explicitly target:

1. Invoking a script called "npm" with a variety of different scenarios
   to mirror the use cases that led to "npm" getting special treatment
   in the first place.  Especially, "npm" as a script in the PATH vs a
   full absolute path to "npm" where it's not the one in the PATH, vs a
   full absolute path to "npm" which is _different_ from the npm in the
   PATH.

2. Executing/spawning a different "node" than the process.execPath,
   either via PATH lookup, or absolute path, both where execPath is in
   the PATH and where it is not.  (Ie, the bug in #38.)

3. Adding a spawn-wrap from within a spawn-wrapped process, and
   verifying that both wraps execute.