Support running one version of node from another
This fixes #38, and also a few other obvious bugs introduced
in the massive refactor of the previous commit.

Instead of always running our shim instead of node, we now shove the
shim in between the user provided node and the rest of the args.

If someone does 'node foo' from a shell script, this will still result
in the shim being called directly, which is intended.  In order to
prevent double-shimming, we must resolve spawn('node',...) calls to a
fully absolute path.  Another approach would be to just avoid adding the
shim if we know that PATH resolution will result in our node shim being
found anyway.