fix: Make munge functions pure
This commit updates the munge functions to no longer update the options object in place but return the updated options. If the options are not modified, it may return the input options directly.

This allows better isolation of the munge functions so they are easier to verify.

As part of this change, the mungers no longer set the custom field `basename` on the options object. It was previously used as an indirect way to pass data between the `shebang` and `node` mungers. It got replaced by a standalone function to retrieve the basename from the file.