fix(win32): handle cases where "node" is quoted
This fixes an issue which was triggered by npm v6.11, where the
"program" referenced by a shimmed shebang is always quoted (whereas
before, it would not be).  Ironically, the code change in npm was to
work around cases where a user puts `"jslint"` in a batch file,
resulting in the wrong %~dp0 being loaded.  The fix triggered the same
behavior in NYC, and the same fix resolves the problem here.

For arcane and mysterious reasons, calling a subroutine in a batch file
causes the %~dp0 variable to be set properly.